# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AnotherViewer WebUI Backend API
  description: |
    REST API for the AnotherViewer WebUI backend (Spring Boot).

    Authentication: all endpoints under /api/v1 require a bearer token
    (Authorization: Bearer <token>) except the publicly permitted ones
    (auth/register, auth/login, auth/status, auth/pair/complete, health,
    metrics, metrics/dashboard, GET site/availability) and non-API
    resources.
  version: 1.1.0
  contact:
    name: AnotherViewer WebUI Team
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/gallery/search:
    get:
      tags: [Gallery]
      operationId: searchGallery
      summary: Search galleries with keyword, category filter, and pagination
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: Search keyword
        - name: category
          in: query
          required: false
          schema:
            type: integer
          description: Category bitmask filter
        - name: sort
          in: query
          required: false
          schema:
            type: integer
            enum: [0, 1, 2, 3]
            default: 0
          description: >-
            Sort order (v1.1+): 0/absent = default order, 1 = posted time desc,
            2 = rating desc, 3 = title asc. Passed through to the site URL as
            f_order.
        - name: pageMin
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Minimum page count (site params f_sp=on&f_spf=N); absent = no bound
        - name: pageMax
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Maximum page count (site params f_sp=on&f_spt=N); absent = no bound
        - name: minRating
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 5
          description: Minimum rating threshold (site params f_sr=on&f_srdd=N); 0/absent = disabled
        - name: searchName
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced search scope (advsearch=1&f_sname=on): match gallery name"
        - name: searchTags
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced search scope (advsearch=1&f_stags=on): match tags"
        - name: searchDesc
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced search scope (advsearch=1&f_sdesc=on): match description"
        - name: searchTorrents
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced search scope (advsearch=1&f_storr=on): match torrent filenames"
        # W3 R4-10 — higher AdvanceSearchTable bits (backendized so the WebUI
        # single filter surface can express the full Android option set).
        - name: searchTorrentsOnly
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced option (advsearch=1&f_sto=on): only show galleries with torrents"
        - name: searchLowPowerTags
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced option (advsearch=1&f_sdt1=on): also match low-power tags"
        - name: searchDownvotedTags
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced option (advsearch=1&f_sdt2=on): also match downvoted tags"
        - name: searchExpunged
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced option (advsearch=1&f_sh=on): include expunged galleries"
        - name: disableLanguageFilter
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced option (advsearch=1&f_sfl=on): disable the default language filter"
        - name: disableUploaderFilter
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced option (advsearch=1&f_sfu=on): disable the default uploader filter"
        - name: disableTagFilter
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: "Advanced option (advsearch=1&f_sft=on): disable the default tag filter"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (0-based)
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Number of results per page
      responses:
        '200':
          description: >-
            Gallery list. When the Gallery Site is circuit-broken, returns
            success=false, empty data, total=0 and cause="EH_UNAVAILABLE"
            (no upstream I/O; see /api/v1/site/availability).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryListResponse'
  /api/v1/gallery/feed:
    get:
      tags: [Gallery]
      operationId: galleryFeed
      summary: >-
        Gallery feed by mode: subscription (watched), popular (what's hot), or
        toplist (EH top lists)
      parameters:
        - name: mode
          in: query
          required: true
          schema:
            type: string
            enum: [subscription, popular, toplist]
          description: >-
            Feed kind. subscription = watched galleries, popular = what's hot,
            toplist = EH top list (first non-empty time slot, yesterday first).
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (0-based); used by subscription/popular only
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Number of results per page (clamped to 1..200)
      responses:
        '200':
          description: >-
            subscription/popular return GalleryListResponse (identical to
            GET /gallery/search); toplist returns TopListResponse. An
            unreachable site surfaces success=false with empty data and
            total 0 (E2E-6, same as /search); when the site is
            circuit-broken the response additionally carries
            cause="EH_UNAVAILABLE".
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/GalleryListResponse'
                  - $ref: '#/components/schemas/TopListResponse'
        '400':
          description: Invalid or missing mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedErrorResponse'
  /api/v1/gallery/{gid}:
    get:
      tags: [Gallery]
      operationId: getGalleryDetail
      summary: Get full detail for a single gallery
      parameters:
        - name: gid
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Gallery ID
      responses:
        '200':
          description: Gallery detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryDetailDto'
        '404':
          description: Gallery not found
  /api/v1/gallery/history:
    get:
      tags: [Gallery]
      operationId: getGalleryHistory
      summary: Get paginated gallery browsing history (via GalleryService)
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Gallery history list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryListResponse'
  /api/v1/gallery/history/{gid}:
    post:
      tags: [Gallery]
      operationId: addToHistory
      summary: Record a gallery visit into history
      parameters:
        - name: gid
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToHistoryRequest'
      responses:
        '200':
          description: Success flag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/v1/gallery/quick-search:
    get:
      tags: [Gallery]
      operationId: getQuickSearches
      summary: List saved quick search presets
      responses:
        '200':
          description: Quick search list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuickSearchListResponse'
    post:
      tags: [Gallery]
      operationId: createQuickSearch
      summary: Create a new quick search preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickSearchDto'
      responses:
        '200':
          description: Created quick search
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuickSearchDto'
  /api/v1/gallery/quick-search/{id}:
    delete:
      tags: [Gallery]
      operationId: deleteQuickSearch
      summary: Delete a quick search preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success flag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================
  # Favorite (remote Gallery Site favorites)
  # ============================================================
...
  /api/v1/cache/gallery/{id}:
    delete:
      tags: [Cache]
      operationId: clearGalleryCache
      summary: Clear all cached images for a specific gallery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Gallery ID
      responses:
        '200':
          description: Success flag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Gallery not found in cache

  # ============================================================
  # Processing
  # ============================================================
  /api/v1/process/gallery/{id}:
    post:
      tags: [Processing]
      operationId: triggerGalleryProcessing
      summary: Trigger AI image enhancement for all pages of a gallery
      description: |
        Enqueues all pages of the gallery for background AI enhancement (e.g. waifu2x upscale/denoise).
        Returns a task ID that can be used to query progress.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Gallery ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessingRequest'
      responses:
        '200':
          description: Processing task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingTaskResponse'
        '404':
          description: Gallery not found
        '409':
          description: Processing already in progress for this gallery
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Bearer token issued by /api/v1/auth/login, /api/v1/auth/register, or /api/v1/auth/pair/complete"
  schemas:
...
    GalleryListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/GalleryItemDto'
        total:
          type: integer
        cause:
          type: string
          enum: [EH_UNAVAILABLE]
          nullable: true
          description: >-
            Optional failure cause; "EH_UNAVAILABLE" when the Gallery Site is
            circuit-broken and the list is empty (success=false, total=0).
            Absent/null on success or plain errors.
    GalleryItemDto:
      type: object
      properties:
        gid:
          type: integer
          format: int64
        token:
          type: string
        galleryUrl:
          type: string
          readOnly: true
        title:
          type: string
          nullable: true
        titleJpn:
          type: string
          nullable: true
        thumb:
          type: string
          nullable: true
        category:
          type: integer
        posted:
          type: string
          nullable: true
        uploader:
          type: string
          nullable: true
        rating:
          type: number
          format: float
        rated:
          type: boolean
        simpleLanguage:
          type: string
          nullable: true
        simpleTags:
          type: array
          items:
            type: string
        thumbWidth:
          type: integer
        thumbHeight:
          type: integer
        pages:
          type: integer
        favoriteSlot:
          type: integer
        favoriteName:
          type: string
          nullable: true
        readProgress:
          type: integer
          minimum: 0
          nullable: true
          description: >-
            0-based last-read page index from the history row (reading
            progress, additive). Absent/null when no progress is known
            (legacy servers or upstream-fetched rows without local history).
    TopListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/TopListFeedItemDto'
        total:
          type: integer
        cause:
          type: string
          enum: [EH_UNAVAILABLE]
          nullable: true
          description: >-
            Optional failure cause; "EH_UNAVAILABLE" when the Gallery Site is
            circuit-broken and the list is empty (success=false, total=0).
            Absent/null on success or plain errors.
    TopListFeedItemDto:
      type: object
      properties:
        gid:
          type: string
          nullable: true
        token:
          type: string
          nullable: true
        tag:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
        href:
          type: string
          nullable: true
    FeedErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
    GalleryDetailDto:
      type: object
      properties:
        gid:
          type: integer
          format: int64
        token:
          type: string
        galleryUrl:
          type: string
          readOnly: true
        title:
          type: string
          nullable: true
        titleJpn:
          type: string
          nullable: true
        thumb:
          type: string
          nullable: true
        category:
          type: integer
        posted:
          type: string
          nullable: true
        uploader:
          type: string
          nullable: true
        rating:
          type: number
          format: float
        rated:
          type: boolean
        simpleLanguage:
          type: string
          nullable: true
        simpleTags:
          type: array
          items:
            type: string
        thumbWidth:
          type: integer
        thumbHeight:
          type: integer
        pages:
          type: integer
        favoriteSlot:
          type: integer
        favoriteName:
          type: string
          nullable: true
        readProgress:
          type: integer
          minimum: 0
          nullable: true
          description: >-
            0-based last-read page index from the history row (reading
            progress, additive). Absent/null when no progress is known
            (legacy servers or upstream-fetched rows without local history).
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagDto'
        imageUrl:
          type: string
          nullable: true
        cause:
          type: string
          enum: [EH_UNAVAILABLE]
          nullable: true
          description: >-
            Optional failure cause; "EH_UNAVAILABLE" when the Gallery Site is
            circuit-broken and the detail could not be enriched from upstream.
            Absent/null otherwise.
    TagDto:
      type: object
      properties:
        namespace:
          type: string
          nullable: true
        tag:
          type: string
    AddToHistoryRequest:
      type: object
      properties:
        token:
          type: string
        title:
          type: string
          nullable: true
        page:
          type: integer
          minimum: 0
          description: >-
            0-based reading position to persist with the history row
            (additive). Optional: when absent the server keeps the stored
            progress (legacy clients push no page); an explicit 0 is written
            as-is and means "re-reading from the start".
    QuickSearchListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuickSearchDto'
    QuickSearchDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        mode:
          type: integer
        category:
          type: integer
        keyword:
          type: string
          nullable: true
        advanceSearch:
          type: integer
        minRating:
          type: integer
        pageFrom:
          type: integer
        pageTo:
          type: integer
        sort:
          type: integer
          enum: [0, 1, 2, 3]
          default: 0
          description: >-
            Sort order persisted with the preset (W3 R4-11; same semantics as
            GET /api/v1/gallery/search `sort`): 0 = default order,
            1 = posted time desc, 2 = rating desc, 3 = title asc. Applied
            when the preset is loaded. Absent on legacy rows (reads as 0).
            REST-only: SyncQuickSearchDto (app schema) intentionally has no
            sort — presets keep it when stored via REST, the app ignores it.

    # ----------------------------------------------------------
    # Favorite
    # ----------------------------------------------------------
...
    ProcessingRequest:
      type: object
      properties:
        type:
          type: string
          enum: [UPSCALE_2X, UPSCALE_4X, DENOISE, DENOISE_UPSCALE]
          default: UPSCALE_2X
          description: Processing type to apply
        outputFormat:
          type: string
          enum: [png, webp, jpg]
          default: png
          description: Output image format
        quality:
          type: integer
          default: 90
          minimum: 1
          maximum: 100
          description: Output quality (for lossy formats)
    ProcessingTaskResponse:
      type: object
      properties:
        taskId:
          type: string
          description: Unique task identifier for status polling
        galleryId:
          type: integer
          format: int64
        totalPages:
          type: integer
          description: Number of pages enqueued for processing
        state:
          type: string
          enum: [PENDING, PROCESSING, DONE, FAILED]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
...